Fix bug in editing existing book
When editing an existing book's isbn to an isbn that already exists for
a different book, the program did not display an error but instead
resumed as normal and set the edited isbn to 'null' for some odd reason.
Fixed by moving the code that relies on the isbn value inside the .then
block.